pyproject.toml
[tool.poetry]
name = "codecogs-equation-editor"
version = "0.1.0"
description = "Telegram Bot to send LaTeX formatted Math Equations."
authors = ["Plastic Cat <not_plastic_cat@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.7"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pylint = "^2.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
